Migrate matcher onto the shared vendored core (FuzzyMatcherCore)
Lineuparr is the canonical generator the core was extracted from, so its
FuzzyMatcher now subclasses FuzzyMatcherCore and deletes the 10 inherited
primitives (normalize_name, calculate_similarity, the callsign ladder,
process_string_for_matching, _has_token_overlap, _length_scaled_threshold,
_trailing_number) plus the dead module-level pattern constants. The country-
detection layer, has_upgrade_quality, _is_group_header, _channel_number_boost,
the cache glue, and the entry points (alias_match/fuzzy_match/match_all_streams)
stay Lineuparr-specific. Decorative helpers are re-exported from the core.

Adopts the core's two intended behavior changes (golden baseline regenerated):
  - East/West stripping is now gated on ignore_regional (was unconditional), so
    a regional_off normalize keeps "HBO West"/"HBO East".
  - process_string keeps "+", so Discovery+/Disney+/Paramount+ stay distinct
    (the agreed 4-of-4 superset).

To preserve Lineuparr's region-filter design (strip bare Pacific/Central/
Mountain/Atlantic for scoring, enforce correctness via the post-match filter on
the original names), the core gained an opt-in strip_bare_region resolved from a
_STRIP_BARE_REGION class attr; Lineuparr sets it True. Stream-Mapparr leaves it
off (keeps bug-066: "Comedy Central" survives). Vendored core hash-pinned in
.github/scripts/core_manifest.json (check_core_parity.py + .gitattributes LF);
CI byte-compiles it and runs the parity gate. Full local suite 430 green.
