+++ .gitattributes
@@ -1,0 +1,5 @@
+# The vendored shared core is sha256-pinned by .github/scripts/check_core_parity.py
+# against .github/scripts/core_manifest.json, so it MUST stay LF on every OS - a CRLF
+# checkout (Windows autocrlf) would change the hash and break the drift gate. Same for
+# any future vendored shared file (e.g. aliases_base.py).
+Lineuparr/matching_core.py text eol=lf
