+++ .github/scripts/core_manifest.json
@@ -1,0 +1,3 @@
+{
+  "matching_core.py": "2ac7ac4cb283306152701bfd977ff2f13700d63d8023fdfa280b1480f5a969dc"
+}
